Move RuntimeProvider into proto
